requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.2
distro==1.9.0
dj-database-url==3.0.0
Django==5.2
django-allauth==65.7.0
git-filter-repo==2.47.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
mysqlclient==2.2.7
openai==1.82.1
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.9.0
WeasyPrint==66.0
python-dotenv==1.2.1
pypdf==6.3.0

# Dependencias para Celery y tareas programadas
celery==5.4.0
redis==5.0.1
django-celery-beat==2.8.1

# Dependencias para Google Cloud Platform Gmail API
google-api-python-client==2.162.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-auth==2.38.0
